Added chests with healing potions
added chests with healing potions

+++ C_Code/LARK_Libraries/Player.h
@@ -35,6 +35,6 @@ bool isASolidBlock(char c);
 void OnPlayerUpdate(Transform* this);
 
 //used for damaging or restoring health
-void ModifyHealth(float* hp);
+void ModifyHealth(int* hp);
 
 #endif
